Remove Home and replace with Manage CMs
