+++ pages/[id].tsx
